+++ pyoxidizer/src/pyembed/pyinterp.rs
@@ -355,6 +355,7 @@ impl<'a> MainPythonInterpreter<'a> {
             pyffi::Py_IgnoreEnvironmentFlag = if config.ignore_python_env { 1 } else { 0 };
             pyffi::Py_InteractiveFlag = if config.interactive { 1 } else { 0 };
             pyffi::Py_InspectFlag = if config.inspect { 1 } else { 0 };
+            pyffi::Py_IsolatedFlag = if config.isolated { 1 } else { 0 };
             pyffi::Py_NoSiteFlag = if config.import_site { 0 } else { 1 };
             pyffi::Py_NoUserSiteDirectory = if config.import_user_site { 0 } else { 1 };
             pyffi::Py_OptimizeFlag = config.opt_level;